Modify dataloading in solvers (#286)
